requirements.txt
Flask==2.3.2
gTTS==2.3.0
moviepy==1.0.3
requests==2.31.0
Pillow==10.0.0
pydub==0.25.1
gunicorn